chore(asm): update the event tracking SDK to the latest spec changes
