+++ ddtrace/ext/user.py
@@ -5,3 +5,4 @@ EMAIL = "usr.email"
 ROLE = "usr.role"
 SCOPE = "usr.scope"
 SESSION_ID = "usr.session_id"
+EXISTS = "usr.exists"
