+++ releasenotes/notes/asm-event-tracking-sdk-d7e8516511c73589.yaml
@@ -1,4 +1,4 @@
 ---
 features:
   - |
-    ASM: Added two new functions for the Application Security Events Tracking API; `track_user_login_event` and `track_custom_event`.
+    ASM: Added three new functions for the Application Security Events Tracking API in the `appsec.trace_utils` module; `track_user_login_success_event`, `appsec.track_user_login_failure_event` and `track_custom_event`.
